# docs/API.yaml
openapi: 3.0.3
info:
  title: Tasky API
  description: |
    REST API for Tasky — Mongolia's most trusted domestic service marketplace.

    ## Authentication
    All endpoints (unless noted) require a Bearer JWT in the `Authorization` header.
    Obtain tokens via the OTP login flow (`/auth/otp/*`).

    ## Pagination
    All list endpoints use **cursor-based pagination**.
    Pass `cursor` (opaque string) and `limit` (default 20, max 100) as query parameters.
    Responses wrap results in `{ data: [], cursor: { next, has_more } }`.

    ## Errors
    All errors return a standard JSON body: `{ code, message, trace_id }`.

    ## Currency
    All monetary amounts are integers in MNT (Mongolian Tugrik).

    ## Idempotency
    Critical state-changing endpoints require an `Idempotency-Key` header.
    Payment and wallet endpoints are documented as post-MVP deferred surfaces.
  version: 1.1.0
  contact:
    name: Tasky Engineering
servers:
  - url: /api/v1
    description: API v1
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    CursorParam:
      name: cursor
      in: query
      description: Opaque cursor for pagination. Pass the `next` value from a previous response.
      schema:
        type: string
    LimitParam:
      name: limit
      in: query
      description: Maximum number of items to return (default 20, max 100).
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    PathId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    Error:
      type: object
      required: [code, message, trace_id]
      properties:
        code:
          type: string
          example: TASK_ALREADY_BOOKED
        message:
          type: string
          example: This task has already been assigned to another tasker.
        trace_id:
          type: string
          example: abc-123-def
    CursorPagination:
      type: object
      required: [next, has_more]
      properties:
        next:
          type: string
          nullable: true
          description: Opaque cursor to fetch the next page. Null if no more results.
          example: eyJpZCI6MTAwfQ==
        has_more:
          type: boolean
          example: true
...
    User:
      type: object
      required: [id, phone, role, status, created_at]
      properties:
        id:
          type: string
          format: uuid
        phone:
          type: string
          example: "+97699001122"
        role:
          type: string
          enum: [CUSTOMER, TASKER, ADMIN]
        status:
          type: string
          enum: [PENDING, VERIFIED, BANNED, SUSPENDED]
        created_at:
          type: string
          format: date-time
...
    VerificationDetail:
      type: object
      required: [id, user_id, status, submitted_at]
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        user_phone:
          type: string
        user_name:
          type: string
        id_card_front_url:
          type: string
          format: uri
        id_card_back_url:
          type: string
          format: uri
        status:
          type: string
          enum: [PENDING, APPROVED, REJECTED]
        admin_notes:
          type: string
          nullable: true
        submitted_at:
          type: string
          format: date-time
        reviewed_at:
          type: string
          format: date-time
          nullable: true

    # ─── Categories ───────────────────────────────────────────
...
  responses:
...
    Unauthorized:
      description: Missing or invalid authentication token.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Insufficient permissions for this action.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Requested resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
paths:
...
  /admin/users:
    get:
      tags: [Admin]
      summary: Search users
      description: Search users by phone number. Admin only.
      operationId: adminSearchUsers
      security:
        - BearerAuth: []
      parameters:
        - name: phone
          in: query
          description: Phone number to search for (partial match supported).
          schema:
            type: string
        - $ref: '#/components/parameters/CursorParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Paginated list of users.
          content:
            application/json:
              schema:
                type: object
                required: [data, cursor]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
                  cursor:
                    $ref: '#/components/schemas/CursorPagination'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/users/{id}/ban:
    post:
      tags: [Admin]
      summary: Ban a user
      description: |
        Bans a user, preventing them from logging in even with a valid JWT.
        Active bookings are not automatically cancelled.
      operationId: adminBanUser
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PathId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason:
                  type: string
                  minLength: 1
      responses:
        '200':
          description: User banned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: User is already banned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /admin/users/{id}/unban:
    post:
      tags: [Admin]
      summary: Unban a user
      description: Removes the ban on a user, restoring their access.
      operationId: adminUnbanUser
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PathId'
      responses:
        '200':
          description: User unbanned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: User is not banned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /admin/verifications/pending:
    get:
      tags: [Admin]
      summary: List pending verifications
      description: Returns paginated list of verification submissions awaiting admin review.
      operationId: adminListPendingVerifications
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CursorParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Paginated list of pending verifications.
          content:
            application/json:
              schema:
                type: object
                required: [data, cursor]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/VerificationDetail'
                  cursor:
                    $ref: '#/components/schemas/CursorPagination'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/verifications/{id}/approve:
    post:
      tags: [Admin]
      summary: Approve verification
      description: Approves a tasker's identity verification. User status changes to VERIFIED.
      operationId: adminApproveVerification
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PathId'
      responses:
        '200':
          description: Verification approved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerificationDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Verification is not in PENDING status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /admin/verifications/{id}/reject:
    post:
      tags: [Admin]
      summary: Reject verification
      description: Rejects a tasker's identity verification with a reason. Tasker can resubmit.
      operationId: adminRejectVerification
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PathId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason:
                  type: string
                  minLength: 1
                  description: Reason for rejection, visible to the tasker.
      responses:
        '200':
          description: Verification rejected.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerificationDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Verification is not in PENDING status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...